Updated api docs, db schema
Schema:
To stay with convention, consider changing “belongsToReview” on your Review_Images table to be something related to foreign id. Example: reviewId or review_id
^ same note but for Spot_Images table
Docs:
Refactor “Create A Spot” end point to be RESTful: We do not want to have /spots/current_user as this breaks convention on what the end point does. We can refactor this to be /spots to just declare that we are posting, or creating something, on the spots table.
Refactor “Add an image to a review based on the Review’s id”. Restful convention would make this shorter as do not need to reference spots as the association between review and images is enough and the connection to spots from reviews is implied. Example: /reviews/:reviewId/review-images
^ same note but for “Edit a Review”. We can just edit the review without having to reference the spot. So we can change: /spots/:spotId/reviews/:reviewId to be /reviews/:reviewId. We can think of this as a PUT being used to edit a review on the review table, and the connection between review and spots is implied and not needed in the route itself. If we were doing this for review images, we may need to reference reviews first and then images because the route may just be something like reviews/:id/images and there is nothing telling us which image table or anything.
Refactor Restful URL for “Delete a Review”
refactor Restful URL for “Edit a booking”
Refactor Restful URL for “Delete a booking”
Refactor Restful URL for “Delete a Review Image”
Add a RESTful url for “Add query filters to get all spots” /spots